+++ base/src/main/java/proguard/classfile/VersionConstants.java
@@ -73,6 +73,8 @@ public class VersionConstants {
   public static final int CLASS_VERSION_23_MINOR = 0;
   public static final int CLASS_VERSION_24_MAJOR = 68;
   public static final int CLASS_VERSION_24_MINOR = 0;
+  public static final int CLASS_VERSION_25_MAJOR = 69;
+  public static final int CLASS_VERSION_25_MINOR = 0;
   public static final int PREVIEW_VERSION_MINOR = 65535;
 
   public static final int CLASS_VERSION_1_0 =
@@ -123,7 +125,8 @@ public class VersionConstants {
       (CLASS_VERSION_23_MAJOR << 16) | CLASS_VERSION_23_MINOR;
   public static final int CLASS_VERSION_24 =
       (CLASS_VERSION_24_MAJOR << 16) | CLASS_VERSION_24_MINOR;
-
+  public static final int CLASS_VERSION_25 =
+      (CLASS_VERSION_25_MAJOR << 16) | CLASS_VERSION_25_MINOR;
   public static final int MAX_SUPPORTED_VERSION =
-      (CLASS_VERSION_24_MAJOR << 16) | PREVIEW_VERSION_MINOR;
+      (CLASS_VERSION_25_MAJOR << 16) | PREVIEW_VERSION_MINOR;
 }